+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import './App.css';
-import { Footer } from './components/Footer';
-import { Header } from './components/Header';
+import { Footer } from './components/footer/Footer';
+import { Header } from './components/header/Header';
 
 function App() {
   return (
