+++ src/components/index.ts
@@ -1,2 +1,5 @@
+export * from './footer';
+export * from './header';
 export * from './heroSlide';
+export * from './movies';
 export * from './thumbnail';
